#139 - Remove 2FA tests and refactor ProfileForm tests for improved clarity and structure
